그룹/스팟 완료

+++ app-public-api/src/main/java/co/kurrant/app/public_api/controller/user/UserController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.UnsupportedEncodingException;
+import java.math.BigInteger;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 
@@ -112,7 +113,7 @@ public class UserController {
 
     // TODO: 추후 백오피스 구현시 삭제
     @PostMapping("/setting/corporation/{corporationId}")
-    public ResponseMessage settingCorporation(HttpServletRequest httpServletRequest, @PathVariable Integer corporationId) {
+    public ResponseMessage settingCorporation(HttpServletRequest httpServletRequest, @PathVariable BigInteger corporationId) {
         userService.settingCorporation(httpServletRequest, corporationId);
         return ResponseMessage.builder()
                 .message("유저 그룹(기업) 설정에 성공하였습니다.")
@@ -121,7 +122,7 @@ public class UserController {
 
     // TODO: 추후 백오피스 구현시 삭제
     @PostMapping("/setting/apartment/{apartmentId}")
-    public ResponseMessage settingApartment(HttpServletRequest httpServletRequest, @PathVariable Integer apartmentId) {
+    public ResponseMessage settingApartment(HttpServletRequest httpServletRequest, @PathVariable BigInteger apartmentId) {
         userService.settingApartment(httpServletRequest, apartmentId);
         return ResponseMessage.builder()
                 .message("유저 그룹(아파트) 설정에 성공하였습니다.")
